+++ src/locale/id.js
@@ -1,0 +1,27 @@
+import dayjs from 'dayjs'
+
+const locale = {
+  name: 'id',
+  weekdays: 'Minggu_Senin_Selasa_Rabu_Kamis_Jumat_Sabtu'.split('_'),
+  months: 'Januari_Februari_Maret_April_Mei_Juni_Juli_Agustus_September_Oktober_November_Desember'.split('_'),
+  relativeTime: {
+    future: 'dalam %s',
+    past: '%s yang lalu',
+    s: 'beberapa detik',
+    m: 'semenit',
+    mm: '%d menit',
+    h: 'sejam',
+    hh: '%d jam',
+    d: 'sehari',
+    dd: '%d hari',
+    M: 'sebulan',
+    MM: '%d bulan',
+    y: 'setahun',
+    yy: '%d tahun'
+  },
+  ordinal: n => `${n}.`
+}
+
+dayjs.locale(locale, null, true)
+
+export default locale