+++ src/plugin/advancedFormat/index.js
@@ -1,13 +1,16 @@
-import { FORMAT_DEFAULT } from '../constant'
+import { FORMAT_DEFAULT } from '../../constant'
 
 export default (o, c, d) => { // locale needed later
   const proto = c.prototype
   const oldFormat = proto.format
-  // eslint-disable-next-line no-nested-ternary
-  d.en.ordinal = number => `${number}[${number === 1 ? 'st' : number === 2 ? 'nd' : number === 3 ? 'rd' : 'th'}]`
+  d.en.ordinal = (number) => {
+    const s = ['th', 'st', 'nd', 'rd']
+    const v = number % 100
+    return `[${number}${(s[(v - 20) % 10] || s[v] || s[0])}]`
+  }
   // extend en locale here
-  proto.format = function (formatStr, localeObject) {
-    const locale = localeObject || this.$locale()
+  proto.format = function (formatStr) {
+    const locale = this.$locale()
     const utils = this.$utils()
     const str = formatStr || FORMAT_DEFAULT
     const result = str.replace(/Q|Do|X|x|k{1,2}|S/g, (match) => {
@@ -26,7 +29,7 @@ export default (o, c, d) => { // locale needed later
           return this.$d.getTime()
       }
     })
-    return oldFormat.bind(this)(result, locale)
+    return oldFormat.bind(this)(result)
   }
 }