+++ test/locale/keys.test.js
@@ -15,15 +15,31 @@ fs.readdirSync(path.join(__dirname, localeDir))
 it('Locale keys', () => {
   L.forEach((l) => {
     const {
-      name, ordinal, weekdays, months
+      name,
+      ordinal,
+      weekdays,
+      months,
+      relativeTime,
+      weekdaysShort,
+      monthsShort,
+      weekdaysMin
     } = l
     expect(name).toEqual(expect.any(String))
     expect(weekdays).toEqual(expect.any(Array))
+
+    if (weekdaysShort) expect(weekdaysShort).toEqual(expect.any(Array))
+    if (monthsShort) expect(monthsShort).toEqual(expect.any(Array))
+    if (weekdaysMin) expect(weekdaysMin).toEqual(expect.any(Array))
+
     expect(months).toEqual(expect.any(Array))
-    if (ordinal) {
-      // function pass date return string or number or null
-      expect(ordinal(1)).toEqual(expect.anything())
-    }
+    // function pass date return string or number or null
+    expect(ordinal(1)).toEqual(expect.anything())
+    expect(ordinal(3)).toEqual(expect.anything())
     expect(dayjs().locale(name).$locale().name).toBe(name)
+    if (relativeTime) {
+      expect(Object.keys(relativeTime).sort()).toEqual(['d', 'dd', 'future', 'h', 'hh', 'm', 'mm', 'M', 'MM',
+        'past', 's', 'y', 'yy']
+        .sort())
+    }
   })
 })